Global Variable Documentation
ven though we wrote it as HTML, we still have to save it with a .txt extension as mentioned above. Commit the defineProps.txt file.
NOTE: Global Variable Documentation for custom steps will only be availale under a Pipeline job that uses it and has run successfully at least once.